backend/requirements.txt
# Database
motor>=3.3.2
pymongo>=4.6.1

# Async HTTP
aiohttp>=3.9.1

# Environment variables
python-dotenv>=1.0.0

# Authentication
pyjwt>=2.8.0
bcrypt>=4.1.2

# Date/Time handling
pytz>=2024.1

# Logging
logging>=0.4.9.6

# Google Calendar API (for future integration)
google-auth>=2.27.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.118.0

# Type hints
typing-extensions>=4.9.0

# Optional but recommended for development
pytest>=8.0.0
pytest-asyncio>=0.23.5
pytest-cov>=4.1.0
black>=24.1.1
flake8>=7.0.0

# API Framework
fastapi>=0.109.2
uvicorn>=0.27.1
pydantic>=2.6.1
pydantic[email]>=2.6.1

# Testing
pytest>=8.0.0
httpx>=0.26.0  # for testing FastAPI
pytest-asyncio>=0.23.5
pytest-cov>=4.1.0

# Development
black>=24.1.1
flake8>=7.0.0